Makefile: only use DESTDIR in install rule.
DESTDIR should only override the overall root of the installation.
Remove $(DESTDIR) reference from definitions of BINDIR and MANDIR.
Prefix the references of BINDIR and MANDIR with DESTDIR in the
install rule. This way PREFIX, BINDIR, and MANDIR can all be set
independently without overriding the installation root of DESTDIR.