ensure multiplication takes place in unsigned long
the cast here seems dobious. the multiplication will take place in
`unsigned int` and the cast only affects the _result_ and not the
multiplication itself. which means that the high 32bits will remain zero
and the shift will thus always produce zero as well.

use UL suffix on the integer constant to ensure that the multiplication
itself happens at `unsigned long`.